Fix comment nesting
